+++ CHANGELOG.md
@@ -5,11 +5,15 @@ This file is used to list changes made in each version of the nugetServer cookbo
 
 0.1.0
 -----
-- [jasonmorgan] - Initial release of nugetServer
+- [jasonmorgan] - Initial release of nuget
 
 1.0.2
 -----
 - [jasonmorgan] - updated nuget client config to bootstrap a nuget repo prior to chocolatey.  Without this the nuget package doesn't install
+
+1.0.6.4
+-------
+- [jasonmorgan] - totally refactored and changed it so it instead provides resources
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.